feat: initial commit with per-request connection
This is a bit hokey, but it achieves the desired result of acquiring a
`PoolConnection` for every routed request, which can then be extracted
in handlers *and* in other middleware.

A mutex has been used to avoid thinking about `Send`/`Sync`, however
only `Send` is necessary so we might be able to get away with a
`RefCell` in future.

There are likely some missing `sqlx` trait implementation that would
help ergonomics, but if more are added then a macro should be used to
implement for each DB!