refactor: add `Transaction` scaffolding
This required making `Ext` connect lazily, since we need a static
`Transaction` which can only be obtained via `Pool::begin`, meaning we
don't know until an extractor is run whether to call `acquire` or
`begin`. This introduced a new runtime error case, if both `Connection`
and `Transaction` extractors are run on the same request.