feat: functional implementation of request-scoped transactions
The `Transaction` extractor will get (or begin) a transaction on the
pool and store it in request extensions. On drop, the transaction is
returned to the extension.

The neat thing is that the `Service`, which injects the request
extension, inspects the response from the inner service and commits the
transaction if the request was successful, otherwise the transaction is
dropped (and so rolled back).

Ultimately this means users would be able to write middleware and
handlers that use the `Transaction` extractor, and would get 'sensible'
per-request transaction behaviour.

There's a lot of tidy-up needed though. In particular, the `Connection`
extractor is probably not worth its weight, and the need for `RwLock` in
`Service::call` kinda sucks.