refactor!: replace DB-specific mods with single `db` mod
The idea is we should have a procedural macro for generating all
implementations for a particular `DB`, so that we just write the
implementations once and stamp them out for each `DB`. For now, we just
implement `postgres` and leave the macro as a chore for later.

BREAKING CHANGE: Trait implementations for `mssql`, `mysql`, and
`sqlite` have been removed.