refactor!: make `Service` infallible
`Service`'s associated `Error` type is now `Infallible`, meaning it can
be used with `axum::Router::layer` directly, without any
`HandleErrorLayer` shenanigans.

This was achieved by using `Error` to convert any `sqlx::Error`s on
commit into a response. The downside of this is that the error message
will be returned to clients, but this will be fixed subsequently.

BREAKING CHANGE: It's no longer possible to layer `axum_sqlx_tx::Layer`
with `HandleErrorLayer`, it should be layered on its own instead.