feat!: introduce `State` to make `Tx` more type-safe
`Tx` now implements `FromRequestParts<State>`, meaning the `Router` must
be provided an instance of `State` in order for it to compile.

`State` can be constructed with `Layer::new`, which now returns a tuple
of `Layer` and `State`. The idea behind constructing both in the same
function is to make it harder to forget to add the `Layer`, however
`Layer::new` is perhaps not the best place for it long-term.

Ultimately this has the desired effect of making `Tx` more type-safe -
applications that attempt to use `Tx` without providing `State` won't
compile, and the API for obtaining `State` makes it harder to forget to
add the `Layer`.

BREAKING CHANGE: `Layer::new` now returns a `(Layer, State)` tuple. This
can be consumed easily by destructuring assignment. `Tx` now requires
`State` to be provided on `Router`s in order for the `Router` to be
usable.